the first part in separating the StandardWellsDense.hpp implementations.
split into commits for easy debugging purpose.
